Added preferenceDemo test, added one null prop and one emit test for dropdown
