+++ src/frontend/shortlist/src/components/preferences/__tests__/PreferenceDemo.test.js
@@ -1,0 +1,15 @@
+import { describe, it, expect, vi } from "vitest";
+import { mount } from "@vue/test-utils";
+import { createTestingPinia } from "@pinia/testing";
+import PreferenceDemo from "../PreferenceDemo.vue";
+
+describe("PreferenceDemo.vue", () => {
+  it("renders with null props", () => {
+    const componentWrapper = mount(PreferenceDemo, {
+      global: {
+        plugins: [createTestingPinia({ createSpy: vi.fn })],
+      },
+    });
+    expect(componentWrapper.exists()).toBe(true);
+  });
+});
