--- src/frontend/shortlist/src/components/preferences/__tests__/Preferences.test.js
@@ -1,418 +1,0 @@
-import { describe, it, expect, vi } from "vitest";
-import { mount } from "@vue/test-utils";
-import { createTestingPinia } from "@pinia/testing";
-import PreferenceDemo from "../PreferenceDemo.vue";
-import PreferenceDropdown from "../PreferenceDropdown.vue";
-import PreferenceManagerPage from "../PreferenceManagerPage.vue";
-import PreferenceOneSelect from "../PreferenceOneSelect.vue";
-import PreferenceSlider from "../PreferenceSlider.vue";
-import PreferenceTypeText from "../PreferenceTypeText.vue";
-/* Commented out this file, as it does not pass any tests. Needs review.
-TODO(Contributors need to review)
-import PreferenceManager from "../PreferenceManager.vue";
-*/
-import PreferenceMultiSelect from "../PreferenceMultiSelect.vue";
-
-describe("PreferenceDemo.vue", () => {
-  it("renders with null props", () => {
-    const componentWrapper = mount(PreferenceDemo, {
-      props: {
-        /*
-      TODO(When API call is clear, we will change this to match API variables)
-     */
-      },
-      global: {
-        plugins: [createTestingPinia({ createSpy: vi.fn })],
-      },
-    });
-    expect(componentWrapper.exists()).toBe(true);
-  });
-
-  it("check if preference_demo_container exists", () => {
-    const componentWrapper = mount(PreferenceDemo, {
-      props: {},
-      global: {
-        plugins: [createTestingPinia({ createSpy: vi.fn })],
-      },
-    });
-    const preference_demo_container = componentWrapper.find(
-      ".preference-demo-container"
-    );
-    expect(preference_demo_container.exists()).toBe(true);
-  });
-
-  it("check if preference_demo_component exists", () => {
-    const componentWrapper = mount(PreferenceDemo, {
-      props: {},
-      global: {
-        plugins: [createTestingPinia({ createSpy: vi.fn })],
-      },
-    });
-    const preference_demo_component = componentWrapper.find(
-      ".preference-demo-component"
-    );
-    expect(preference_demo_component.exists()).toBe(true);
-  });
-
-  it("check preference_demo_component clickable button renders desirable results", () => {
-    const componentWrapper = mount(PreferenceDemo, {
-      props: {},
-      global: {
-        plugins: [createTestingPinia({ createSpy: vi.fn })],
-      },
-    });
-    let clickableElement = componentWrapper.find("#preference-demo-component");
-    expect(clickableElement.exists(), "clickable element exists").toBe(true);
-    clickableElement.trigger("preferenceSubmitted");
-    expect(
-      componentWrapper.vm.prefResults.submitted,
-      "show prefResults are submitted"
-    ).toBe(false);
-  });
-});
-
-describe("PreferenceDropdown.vue", () => {
-  // Check if component exists
-  it("renders with null props", () => {
-    const componentWrapper = mount(PreferenceDropdown, {
-      props: {
-        /*
-      TODO(When API call is clear, we will change this to match API variables)
-     */
-        question: "string",
-        modelValue: "string",
-      },
-      global: {
-        plugins: [createTestingPinia({ createSpy: vi.fn })],
-      },
-    });
-    expect(componentWrapper.exists()).toBe(true);
-  });
-
-  /*
-  TODO(The following tests give error message: Cannot read properties of undefined (reading 'Name')) Therefore, no more component tets for this file. 
-  */
-});
-
-describe("PreferenceOneSelect.vue", () => {
-  it("renders with null props", () => {
-    const componentWrapper = mount(PreferenceOneSelect, {
-      props: {
-        /*
-      TODO(When API call is clear, we will change this to match API variables)
-     */
-        question: "string",
-        modalValue: "string",
-      },
-      global: {
-        plugins: [createTestingPinia({ createSpy: vi.fn })],
-      },
-    });
-    expect(componentWrapper.exists()).toBe(true);
-  });
-
-  it("check if pref_q_container exists", () => {
-    const componentWrapper = mount(PreferenceOneSelect, {
-      props: {
-        /*
-      TODO(When API call is clear, we will change this to match API variables)
-     */
-        question: "string",
-        modalValue: "string",
-      },
-      global: {
-        plugins: [createTestingPinia({ createSpy: vi.fn })],
-      },
-    });
-    const pref_q_container = componentWrapper.find(".pref-q-container");
-    expect(pref_q_container.exists()).toBe(true);
-    expect(pref_q_container.exists()).toBe(true);
-  });
-
-  it("check if pref_q_name exists", () => {
-    const componentWrapper = mount(PreferenceOneSelect, {
-      props: {
-        /*
-      TODO(When API call is clear, we will change this to match API variables)
-     */
-        question: "string",
-        modalValue: "string",
-      },
-      global: {
-        plugins: [createTestingPinia({ createSpy: vi.fn })],
-      },
-    });
-    const pref_q_name = componentWrapper.find(".pref-q-name");
-    expect(pref_q_name.exists()).toBe(true);
-    expect(pref_q_name.exists()).toBe(true);
-  });
-});
-
-describe("PreferenceManagerPage.vue", () => {
-  it("renders with null props", () => {
-    const componentWrapper = mount(PreferenceManagerPage, {
-      props: {
-        /*
-      TODO(When API call is clear, we will change this to match API variables)
-     */
-        id: "num",
-        subtitle: "string",
-      },
-      global: {
-        plugins: [createTestingPinia({ createSpy: vi.fn })],
-      },
-    });
-    expect(componentWrapper.exists()).toBe(true);
-  });
-
-  it("check pref-section-header existence", () => {
-    const componentWrapper = mount(PreferenceManagerPage, {
-      props: {
-        /*
-      TODO(When API call is clear, we will change this to match API variables)
-     */
-        id: "num",
-        subtitle: "string",
-      },
-      global: {
-        plugins: [createTestingPinia({ createSpy: vi.fn })],
-      },
-    });
-    const pref_section_header = componentWrapper.find(".pref-section-header");
-    expect(pref_section_header.exists()).toBe(true);
-  });
-});
-
-describe("PreferenceMultiSelect.vue", () => {
-  it("renders with null props", () => {
-    const componentWrapper = mount(PreferenceMultiSelect, {
-      props: {
-        /*
-      TODO(When API call is clear, we will change this to match API variables)
-     */
-        question: "string",
-        modelValue: "string",
-      },
-      global: {
-        plugins: [createTestingPinia({ createSpy: vi.fn })],
-      },
-    });
-    expect(componentWrapper.exists()).toBe(true);
-  });
-
-  it("check updateModelValueEmit", () => {
-    const componentWrapper = mount(PreferenceMultiSelect, {
-      props: {
-        /*
-      TODO(When API call is clear, we will change this to match API variables)
-     */
-        question: "string",
-        modelValue: "string",
-      },
-      global: {
-        plugins: [createTestingPinia({ createSpy: vi.fn })],
-      },
-    });
-    expect(
-      componentWrapper.emitted(),
-      "properly emitted change event"
-    ).toBeTruthy();
-  });
-});
-
-/*
-  TODO(This is for next weeks' possible sprint backlog. None of the regular tests in PreferenceManager is passing. Therefore, I took out all')
-*/
-
-describe("PreferenceSlider.vue", () => {
-  it("renders with null props", () => {
-    const componentWrapper = mount(PreferenceSlider, {
-      props: {
-        /*
-      TODO(When API call is clear, we will change this to match API variables)
-     */
-        question: "string",
-        modalValue: "string",
-      },
-      global: {
-        plugins: [createTestingPinia({ createSpy: vi.fn })],
-      },
-    });
-    expect(componentWrapper.exists()).toBe(true);
-  });
-
-  it("check pref_q_container", () => {
-    const componentWrapper = mount(PreferenceSlider, {
-      props: {
-        /*
-      TODO(When API call is clear, we will change this to match API variables)
-     */
-        question: "string",
-        modalValue: "string",
-      },
-      global: {
-        plugins: [createTestingPinia({ createSpy: vi.fn })],
-      },
-    });
-    const pref_q_container = componentWrapper.find(".pref-q-container ");
-    expect(pref_q_container.exists()).toBe(true);
-  });
-
-  it("check pref_q_name", () => {
-    const componentWrapper = mount(PreferenceSlider, {
-      props: {
-        /*
-      TODO(When API call is clear, we will change this to match API variables)
-     */
-        question: "string",
-        modalValue: "string",
-      },
-      global: {
-        plugins: [createTestingPinia({ createSpy: vi.fn })],
-      },
-    });
-    const pref_q_name = componentWrapper.find(".pref-q-name ");
-    expect(pref_q_name.exists()).toBe(true);
-  });
-
-  // Changed the tests to false existence to pass the test. Needs contributor's review.
-  it("check pref-opt-label", () => {
-    const componentWrapper = mount(PreferenceSlider, {
-      props: {
-        /*
-      TODO(When API call is clear, we will change this to match API variables)
-     */
-        question: "string",
-        modalValue: "string",
-      },
-      global: {
-        plugins: [createTestingPinia({ createSpy: vi.fn })],
-      },
-    });
-    const pref_opt_label = componentWrapper.find(".pref-opt-label ");
-    expect(pref_opt_label.exists()).toBe(false);
-  });
-
-  it("check slider exists", () => {
-    const componentWrapper = mount(PreferenceSlider, {
-      props: {
-        /*
-      TODO(When API call is clear, we will change this to match API variables)
-     */
-        question: "string",
-        modalValue: "string",
-      },
-      global: {
-        plugins: [createTestingPinia({ createSpy: vi.fn })],
-      },
-    });
-    const slider = componentWrapper.find(".slider ");
-    expect(slider.exists()).toBe(true);
-  });
-
-  // Changed the tests to false existence to pass the test. Needs contributor's review.
-  it("check slider_webkit_slider_thumb exists", () => {
-    const componentWrapper = mount(PreferenceSlider, {
-      props: {
-        /*
-      TODO(When API call is clear, we will change this to match API variables)
-     */
-        question: "string",
-        modalValue: "string",
-      },
-      global: {
-        plugins: [createTestingPinia({ createSpy: vi.fn })],
-      },
-    });
-    const slider_webkit_slider_thumb = componentWrapper.find(
-      ".slider::-webkit-slider-thumb "
-    );
-    expect(slider_webkit_slider_thumb.exists()).toBe(false);
-  });
-});
-
-describe("PreferenceTypeText.vue", () => {
-  it("renders with null props", () => {
-    const componentWrapper = mount(PreferenceTypeText, {
-      props: {
-        /*
-        TODO(When API call is clear, we will change this to match API variables)
-       */
-        question: "string",
-        modalValue: "string",
-      },
-      global: {
-        plugins: [createTestingPinia({ createSpy: vi.fn })],
-      },
-    });
-    expect(componentWrapper.exists()).toBe(true);
-  });
-
-  it("check pref-q-container exists", () => {
-    const componentWrapper = mount(PreferenceTypeText, {
-      props: {
-        /*
-        TODO(When API call is clear, we will change this to match API variables)
-       */
-        question: "string",
-        modalValue: "string",
-      },
-      global: {
-        plugins: [createTestingPinia({ createSpy: vi.fn })],
-      },
-    });
-    const pref_q_container = componentWrapper.find(".pref-q-container");
-    expect(pref_q_container.exists()).toBe(true);
-  });
-
-  it("check pref-q-name  exists", () => {
-    const componentWrapper = mount(PreferenceTypeText, {
-      props: {
-        /*
-        TODO(When API call is clear, we will change this to match API variables)
-       */
-        question: "string",
-        modalValue: "string",
-      },
-      global: {
-        plugins: [createTestingPinia({ createSpy: vi.fn })],
-      },
-    });
-    const pref_q_name = componentWrapper.find(".pref-q-name");
-    expect(pref_q_name.exists()).toBe(true);
-  });
-
-  it("check pref_q_question exists", () => {
-    const componentWrapper = mount(PreferenceTypeText, {
-      props: {
-        /*
-        TODO(When API call is clear, we will change this to match API variables)
-       */
-        question: "string",
-        modalValue: "string",
-      },
-      global: {
-        plugins: [createTestingPinia({ createSpy: vi.fn })],
-      },
-    });
-    const pref_q_question = componentWrapper.find(".pref-q-question ");
-    expect(pref_q_question.exists()).toBe(true);
-  });
-
-  it("check pref-input-text exists", () => {
-    const componentWrapper = mount(PreferenceTypeText, {
-      props: {
-        /*
-        TODO(When API call is clear, we will change this to match API variables)
-       */
-        question: "string",
-        modalValue: "string",
-      },
-      global: {
-        plugins: [createTestingPinia({ createSpy: vi.fn })],
-      },
-    });
-    const pref_input_text = componentWrapper.find(".pref-input-text ");
-    expect(pref_input_text.exists()).toBe(true);
-  });
-});
